add postgreSql connection
