+++ .gitignore
@@ -15,6 +15,9 @@
 *.log
 temp_file
 
+#local settings
+.vscode/
+
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
