+++ cmd/agent/main.go
@@ -5,13 +5,13 @@ import (
 
 	"github.com/and161185/metrics-alerting/internal/client"
 	"github.com/and161185/metrics-alerting/internal/config"
-	"github.com/and161185/metrics-alerting/storage"
+	"github.com/and161185/metrics-alerting/storage/inmemory"
 )
 
 func main() {
 
 	config := config.NewClientConfig()
-	storage := storage.NewMemStorage()
+	storage := inmemory.NewMemStorage()
 	clnt := client.NewClient(storage, config)
 
 	if err := clnt.Run(); err != nil {
