+++ cmd/server/main.go
@@ -1,20 +1,22 @@
 package main
 
 import (
-	"log"
-
 	"github.com/and161185/metrics-alerting/internal/config"
 	"github.com/and161185/metrics-alerting/internal/server"
-	"github.com/and161185/metrics-alerting/storage"
+	"github.com/and161185/metrics-alerting/storage/postgres"
 )
 
 func main() {
 
 	config := config.NewServerConfig()
-	storage := storage.NewMemStorage()
+	//storage := inmemory.NewMemStorage()
+	storage, err := postgres.NewPostgresStorage(config.DatabaseDsn)
+	if err != nil {
+		config.Logger.Fatal(err)
+	}
 
 	srv := server.NewServer(storage, config)
 	if err := srv.Run(); err != nil {
-		log.Fatal(err)
+		config.Logger.Fatal(err)
 	}
 }
