+++ internal/errs/errs.go
@@ -1,0 +1,5 @@
+package errs
+
+import "errors"
+
+var ErrMetricNotFound = errors.New("metric not found")
